+++ CHANGELOG.md
@@ -7,6 +7,7 @@ unreleased
 - Fix error with `toLowerCase` on error reporting
 - Update braintree-web to v3.42.0
 - Update @braintree/asset-loader to v0.2.1
+- Fix issue where 3ds modal may not get cleaned up during teardown (#463)
 
 1.14.1
 ------